% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_packages_for_renv_install}
\alias{extract_packages_for_renv_install}
\title{Extract vector of packages from renv lockfile for \link[renv:install]{renv::install}}
\usage{
extract_packages_for_renv_install(renv_lock = "renv.lock")
}
\arguments{
\item{renv_lock}{[\link{character}] Path to renv.lock.}
}
\value{
Character vector of packages.
}
\description{
Extract vector of packages from renv lockfile for \link[renv:install]{renv::install}
}
\keyword{internal}
