% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_rspm_packages}
\alias{extract_rspm_packages}
\title{Extract vector of CRAN packages from renv lockfile}
\usage{
extract_rspm_packages(renv_lock)
}
\arguments{
\item{renv_lock}{Character. Path to renv.lock.}
}
\value{
Character vector of packages.
}
\description{
Extract vector of CRAN packages from renv lockfile
}
\keyword{internal}
