% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_options}
\alias{set_options}
\title{Set options}
\usage{
set_options(
  repos = c(RSPM = "https://packagemanager.rstudio.com/all/__linux__/focal/latest"),
  pkgType = "binary"
)
}
\arguments{
\item{repos}{[\link{character}]}

\item{pkgType}{[\link{character}]}
}
\value{

}
\description{
Currently:
\itemize{
\item RSPM (RStudio Package Manager) repo for Ubuntu focal (20.04)
\item Binary package type
}
}
\examples{
set_options()
}
